// packages/personal-note/src/types.ts
/* (c) Copyright Frontify Ltd., all rights reserved. */

export enum NoteStyle {
    Card = 'Card',
}

export enum NoteBorderRadius {
    None = 'None',
    Small = 'Small',
    Medium = 'Medium',
    Large = 'Large',
}

export enum NoteBorderStyle {
    Solid = 'Solid',
    Dotted = 'Dotted',
    Dashed = 'Dashed',
}